Extract LUCI Analysis client into script/lib/luci_analysis.py
